+++ NEWS.md
@@ -9,6 +9,7 @@
 * Deprecate `id()` for creating key for the consistence of tidy selectors, and will be defunct until end of October, 2019.
 * Continue with deprecation of `fill_na()` in favour of `fill_gaps()`, and will be defunct in the next release.
 * Start with deprecation of `pull_interval()` in favour of `interval_pull()`, and will issue an error in the next release.
+* Deprecate `as.tsibble()` in favour of `as_tsibble()`. The warning is displayed once per session.
 
 ## Improvements